+++ app/(user)/quiz/[id]/QuizClient.tsx
